refactor:Search 컴포넌트 생성, main 코드 리팩토링
