+++ package.json
@@ -10,10 +10,12 @@
   },
   "dependencies": {
     "@tanstack/react-query": "^5.56.2",
+    "@types/swiper": "^6.0.0",
     "axios": "^1.7.7",
     "next": "14.2.7",
     "react": "^18",
     "react-dom": "^18",
+    "swiper": "^11.1.14",
     "tailwind-merge": "^2.5.2"
   },
   "devDependencies": {
