+++ src/components/Main/MostReview.tsx
@@ -1,0 +1,11 @@
+function MostReview() {
+  return (
+    <div className="mt-[93px] flex sm:mt-[142px] 2xl:mt-[158px]">
+      <div className="mx-auto w-[91%] max-w-[1200px]">
+        <p className="text-2lg font-bold sm:text-3xl">🔥 인기 체험</p>
+      </div>
+    </div>
+  );
+}
+
+export default MostReview;
